refactor: rename cpp17 namespace into stdext
This is a breaking change, but can be easily fixed by providing namespace alias.
