+++ bench/formatter/json.cpp
@@ -2,7 +2,7 @@
 
 #include <blackhole/attribute.hpp>
 #include <blackhole/attributes.hpp>
-#include <blackhole/cpp17/string_view.hpp>
+#include <blackhole/stdext/string_view.hpp>
 #include <blackhole/extensions/writer.hpp>
 #include <blackhole/formatter.hpp>
 #include <blackhole/formatter/json.hpp>
