+++ include/blackhole/attribute.hpp
@@ -8,7 +8,7 @@
 #include <boost/mpl/contains.hpp>
 #include <boost/mpl/vector.hpp>
 
-#include "blackhole/cpp17/string_view.hpp"
+#include "blackhole/stdext/string_view.hpp"
 
 namespace blackhole {
 inline namespace v1 {
