+++ include/blackhole/message.hpp
@@ -2,7 +2,7 @@
 
 #include <functional>
 
-#include "blackhole/cpp17/string_view.hpp"
+#include "blackhole/stdext/string_view.hpp"
 
 namespace blackhole {
 inline namespace v1 {
