+++ src/procname.cpp
@@ -10,7 +10,7 @@
 
 #include <cstring>
 
-#include "blackhole/cpp17/string_view.hpp"
+#include "blackhole/stdext/string_view.hpp"
 
 namespace blackhole {
 inline namespace v1 {
@@ -18,26 +18,26 @@ namespace detail {
 
 namespace {
 
-auto procname(pid_t pid) -> cpp17::string_view {
+auto procname(pid_t pid) -> stdext::string_view {
 #ifdef __linux__
     (void)pid;
-    return cpp17::string_view(program_invocation_short_name, ::strlen(program_invocation_short_name));
+    return stdext::string_view(program_invocation_short_name, ::strlen(program_invocation_short_name));
 #elif __APPLE__
     static char path[PROC_PIDPATHINFO_MAXSIZE];
 
     if (::proc_name(pid, path, sizeof(path)) > 0) {
-        return cpp17::string_view(path, ::strlen(path));
+        return stdext::string_view(path, ::strlen(path));
     } else {
         const auto nwritten = ::snprintf(path, sizeof(path), "%d", pid);
-        return cpp17::string_view(path, static_cast<std::size_t>(nwritten));
+        return stdext::string_view(path, static_cast<std::size_t>(nwritten));
     }
 #endif
 }
 
 }  // namespace
 
-auto procname() -> cpp17::string_view {
-    static const cpp17::string_view name = procname(::getpid());
+auto procname() -> stdext::string_view {
+    static const stdext::string_view name = procname(::getpid());
     return name;
 }
