+++ src/sink/socket/tcp.cpp
@@ -6,7 +6,7 @@
 
 #include "blackhole/config/node.hpp"
 #include "blackhole/config/option.hpp"
-#include "blackhole/cpp17/string_view.hpp"
+#include "blackhole/stdext/string_view.hpp"
 #include "blackhole/extensions/format.hpp"
 #include "blackhole/sink/socket/tcp.hpp"
